requirements.txt
discord.py
python-dotenv
spotipy
yt-dlp
youtube-search-python
